% Generated by roxygen2 (4.0.2): do not edit by hand
\name{counts.for.region}
\alias{counts.for.region}
\title{Test fucntion}
\usage{
counts.for.region(region)
}
\arguments{
\item{region:}{The region of interest, matching a backpage domain}
}
\value{
some data from counts dataframe
}
\description{
Test fucntion
}

