% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffindiff}
\alias{diffindiff}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
diffindiff(target.region, comparison.region.set, event.date, logged = FALSE,
  normalize = FALSE)
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{event.date:}{a YYYY-MM-DD date string for the actual event date}

\item{logged:}{A boolean as to whether to perform a log transform on the data first}
}
\value{
A list with dd, main_diff, comparison_diff, pre, and post values
}
\description{
Take region names and a comparison date and does diff-in-diff
}

