% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/diffindiff_data.R
\name{diffindiff_data}
\alias{diffindiff_data}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
diffindiff_data(target.region, comparison.region.set, event.date,
  logged = FALSE, normalize = FALSE, input_data, date.var = "monthdate",
  group.var = "region", var.of.interest = "counts")
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{event.date:}{a YYYY-MM-DD date string for the actual event date}

\item{logged:}{A boolean as to whether to perform a log transform on the data first}

\item{input_data:}{The dataframe to use for diff-in-diff - needs to have}

\item{date.var:}{String name for the date variable in the input_data dataframe.}

\item{group.var:}{String name for the group (i.e. city name) variable in the input_data dataframe}

\item{var.of.interest:}{String name for the variable of interest (e.g. ad counts) in the input_data dataframe}
}
\description{
Take region names and a comparison date and does diff-in-diff
}
\examples{
diffindiff_data()
}
\keyword{max}

