% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/diffindiff_wrapper.R
\name{diffindiff_wrapper}
\alias{diffindiff_wrapper}
\title{A wrapper for the diffindiff function. This is needed because when passing arguements from python they are converted to a list. This
This will convert them back}
\usage{
diffindiff_wrapper(target.region, comparison.region.set, event.date, input_data)
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{event.date:}{a YYYY-MM-DD date string for the actual event date}

\item{input_data:}{The dataframe to use for diff-in-diff - needs to have}
}
\description{
A wrapper for the diffindiff function. This is needed because when passing arguements from python they are converted to a list. This
This will convert them back
}
\examples{
diffindiff_wrapper()
}
\keyword{diffindiff_wrapper}

