% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_columns}
\alias{get_columns}
\title{Tabulate the unique values of a given column}
\usage{
get_columns(data)
}
\arguments{
\item{data:}{The data frame that will be analyzed}
}
\value{
A list of the columns of the data frame
}
\description{
Tabulate the unique values of a given column
}

