% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_features}
\alias{get_features}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
get_features(target.region, num.matches = 4)
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{num.matches:}{the number of matches to return. Defaults to 4.}
}
\value{
A data frame with:
- matched regions as rows
- features as columns
- score as a final column measuring similarity
}
\description{
Take region names and a comparison date and does diff-in-diff
}

