% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_features_data}
\alias{get_features_data}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
get_features_data(target.region = "dc", num.matches = 4,
  region.var = "region", comparison.vars, input_data)
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{num.matches:}{the number of matches to return. Defaults to 4.}
}
\value{
A data frame with:
- matched regions as rows
- features as columns
- score as a final column measuring similarity
}
\description{
Take region names and a comparison date and does diff-in-diff
}

