% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_unit_names}
\alias{get_unit_names}
\title{Tabulate the unique values of a given column}
\usage{
get_unit_names(data, column)
}
\arguments{
\item{data:}{The data frame that will be analyzed}

\item{column:}{The column that will be used}
}
\value{
A list of unique values from the column
}
\description{
Tabulate the unique values of a given column
}

