% Generated by roxygen2 (4.0.2): do not edit by hand
\name{log_diagnostic}
\alias{log_diagnostic}
\title{Determine whether to apply logs to a data column}
\usage{
log_diagnostic(df, column)
}
\arguments{
\item{df:}{The data frame that will be analyzed}

\item{column:}{The column that will be used}
}
\value{
A list with values:
log: a boolean which is TRUE if log is better, and FALSE if not
}
\description{
Determine whether to apply logs to a data column
}

