% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quandl_test}
\alias{quandl_test}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
quandl_test(start_date = "2012-01-01", end_date = "")
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{event.date:}{a YYYY-MM-DD date string for the actual event date}
}
\value{
A list with dd, main_diff, comparison_diff, pre, and post values
}
\description{
Take region names and a comparison date and does diff-in-diff
}

