% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/show_contents.R
\name{show_contents}
\alias{show_contents}
\title{A test function so we can examine how content looks when it's passed to opencpu}
\usage{
show_contents(value)
}
\arguments{
\item{value:}{the value passed}
}
\description{
A test function so we can examine how content looks when it's passed to opencpu
}
\examples{
show_contents()
}
\keyword{show_contents}

