% Generated by roxygen2 (4.0.2): do not edit by hand
\name{store_csv}
\alias{store_csv}
\title{Take region names and lookup series from a panel}
\usage{
store_csv(data, t = "asfs")
}
\arguments{
\item{data:}{a csv file which is written to disk}
}
\value{
The dimensions of the loaded dataframe
}
\description{
Take region names and lookup series from a panel
}

