% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/twolines_data.R
\name{twolines_data}
\alias{twolines_data}
\title{Take region names and lookup series from a panel}
\usage{
twolines_data(target.region, comparison.region.set, data, date.var = "date",
  var.of.interest = "counts", group.var = "region")
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{data:}{The data frame to split up - needs columns 'region', 'MonthDate', and 'counts'}
}
\description{
Take region names and lookup series from a panel
}
\examples{
twolines_data()
}
\keyword{max}

