% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffindiff_data}
\alias{diffindiff_data}
\title{Take region names and a comparison date and does diff-in-diff}
\usage{
diffindiff_data(target.region, comparison.region.set, event.date,
  logged = FALSE, input_data, date.var = "MonthDate", group.var = "group",
  var.of.interest = "counts", region.var = "region")
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{event.date:}{a YYYY-MM-DD date string for the actual event date}

\item{logged:}{A boolean as to whether to perform a log transform on the data first}

\item{input_data:}{The dataframe to use for diff-in-diff - needs to have}

\item{date.var:}{String name for the date variable in the input_data dataframe.}

\item{group.var:}{String name for the group (i.e. city name) variable in the input_data dataframe}

\item{var.of.interest:}{String name for the variable of interest (e.g. ad counts) in the input_data dataframe}
}
\value{
A list with dd, main_diff, comparison_diff, pre, and post values
}
\description{
Take region names and a comparison date and does diff-in-diff
}

