% Generated by roxygen2 (4.0.2): do not edit by hand
\name{twolines_data}
\alias{twolines_data}
\title{Take region names and lookup series from a panel}
\usage{
twolines_data(target.region, comparison.region.set, data,
  date.var = "MonthDate", group.var = "group", var.of.interest = "counts",
  region.var = "region")
}
\arguments{
\item{target.region:}{The region of interest, matching a backpage domain}

\item{comparison.region.set:}{A set of regions (e.g. c('nova','abilene')) to compare to}

\item{data:}{The data frame to split up - needs columns 'region', 'MonthDate', and 'counts'}
}
\value{
A dataframe with 'MonthDate', target.varname and 'comparison' columns
}
\description{
Take region names and lookup series from a panel
}

